updated amount/balance type to accept float as well

+++ src/user/dto/balance.response.dto.ts
@@ -1,7 +1,7 @@
-import { Field, Int, ObjectType } from '@nestjs/graphql';
+import { Field, Float, Int, ObjectType } from "@nestjs/graphql";
 
 @ObjectType()
 export class BalanceResponseDto {
-  @Field(() => Int)
+  @Field(() => Float)
   balance: number;
 }
